Size as long

+++ TAG.Networking.DockerRegistry/Model/IImageLayer.cs
@@ -3,7 +3,7 @@
 	public interface IImageLayer
 	{
 		public string MediaType { get; }
-		public int Size { get; }
+		public long Size { get; }
 		public HashDigest Digest { get; }
 		public string[] Urls { get; }
 	}
